templated ofxUISlider, types int, float, double

+++ src/ofxUIUtils.h
@@ -38,7 +38,7 @@ string ofxUIToString(const T& value, int precision){
 	return out.str();
 }
 
-static float ofxUIMap(float value, float inputMin, float inputMax, float outputMin, float outputMax, bool clamp)
+static float ofxUIMap(double value, double inputMin, double inputMax, double outputMin, double outputMax, bool clamp)
 {    
 	if (fabs(inputMin - inputMax) < FLT_EPSILON)
     {
